refactor: Rename default to env to be more consistent with /app config
